backend/requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.2
pydantic-settings==2.0.3
python-multipart==0.0.6
python-dotenv==1.0.0
pypdf==3.17.1
python-docx==0.8.11
pytesseract==0.3.10
pdf2image==1.17.0
pillow==10.2.0
spacy==3.7.2
dateparser==1.1.8
opencv-python-headless==4.8.1.78
anthropic==0.17.0  
sentence-transformers==2.2.2
faiss-cpu==1.7.4
motor==3.3.2
pymongo==4.6.1
redis==5.0.1

prometheus-fastapi-instrumentator==6.1.0
python-json-logger==2.0.7

pytest==7.4.3
httpx==0.25.2